+++ Sources/KeyboardKit/Images/Image+KeyboardKit.swift
@@ -1,5 +1,5 @@
 //
-//  Image+Button.swift
+//  Image+KeyboardKit.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-03-11.