+++ Sources/KeyboardKit/Layout/KeyboardLayoutProvider.swift
@@ -28,10 +28,15 @@ import Foundation
  is ready, just replace the controller service with your own
  implementation to make the library use it instead.
  */
-public protocol KeyboardLayoutProvider: AnyObject, InputSetProviderBased {
+public protocol KeyboardLayoutProvider: AnyObject {
     
     /**
      The layout keyboard to use for a given keyboard context.
      */
     func keyboardLayout(for context: KeyboardContext) -> KeyboardLayout
+    
+    @available(*, deprecated, message: "Use input sets directly instead.")
+    func register(inputSetProvider: InputSetProvider)
 }
+
+