+++ Sources/KeyboardKit/_Deprecated/InputSetProvider.swift
@@ -27,11 +27,8 @@ import Foundation
  new implementation from scratch. When you're implementation
  is ready, just replace the controller service with your own
  implementation to make the library use it instead.
- 
- `v8.0` - This protocol and all implementation types will be
- replaced by just providing a couple of `InputSet` values to
- the layout provider.
  */
+@available(*, deprecated, message: "Use input sets directly instead.")
 public protocol InputSetProvider: AnyObject {
     
     /**